Update rating data type

+++ lib/models/entry.dart
@@ -1,7 +1,7 @@
 class Entry {
   final String title;
   final String description;
-  final double rating;
+  final int rating;
   final DateTime dateTime;
 
   const Entry({this.title, this.description, this.rating, this.dateTime});
